refactor(console): move ConsoleIO-related classes to console package
