--- src/main/kotlin/lotto/ui/ConsoleIoInterface.kt
@@ -1,6 +1,0 @@
-package lotto.ui
-
-interface ConsoleIoInterface : InputReader, OutputReader {
-    override fun read(): String
-    override fun print(message: String)
-}
