+++ src/main/kotlin/lotto/ui/console/ConsoleIOInterface.kt
@@ -1,0 +1,6 @@
+package lotto.ui.console
+
+interface ConsoleIOInterface : InputReader, OutputReader {
+    override fun read(): String
+    override fun print(message: String)
+}
