+++ src/main/kotlin/lotto/ui/console/InputReader.kt
@@ -1,4 +1,4 @@
-package lotto.ui
+package lotto.ui.console
 
 interface InputReader {
     fun read():String
