+++ src/main/kotlin/lotto/ui/console/OutputReader.kt
@@ -1,4 +1,4 @@
-package lotto.ui
+package lotto.ui.console
 
 interface OutputReader {
     fun print(message:String)
